requirements.txt
web3
py-solc
flask
gunicorn